fix: add file existence check before reading script in GithubWorkflowPipeline
Updated the GithubWorkflowPipeline to include a check for the existence of the script file before attempting to read it. This prevents potential errors when the file does not exist. Additionally, updated GitHub Actions workflow snapshots to use the latest versions of actions and dependencies for improved compatibility.

+++ packages/construct/awscdk/github-pipeline/src/workflow.ts
@@ -282,14 +282,19 @@ export class GithubWorkflowPipeline extends ghpipelines.GitHubWorkflow {
 		const jobOutputs = Object.entries(assetHashMap).map(
 			([assetHash, jobName]) => {
 				const scriptPath = path.posix.join(outDir, `publish-${jobName}-step.sh`)
-				const lines = fs
-					.readFileSync(scriptPath, { encoding: 'utf-8' })
-					.toString()
-					.split('\n')
 				const assetNum = jobName.slice(jobName.lastIndexOf('t') + 1)
 				const assetOutput = `asset-hash${assetNum}`
-				lines.push(`echo '${assetOutput}=${assetHash}' >> $GITHUB_OUTPUT`)
-				fs.writeFileSync(scriptPath, lines.join('\n'), { encoding: 'utf-8' })
+
+				// Check if file exists before trying to read it
+				if (fs.existsSync(scriptPath)) {
+					const lines = fs
+						.readFileSync(scriptPath, { encoding: 'utf-8' })
+						.toString()
+						.split('\n')
+					lines.push(`echo '${assetOutput}=${assetHash}' >> $GITHUB_OUTPUT`)
+					fs.writeFileSync(scriptPath, lines.join('\n'), { encoding: 'utf-8' })
+				}
+
 				return [assetOutput, jobName]
 			},
 		)
